populate_sdk_base: Add SDKIMAGE_INSTALL_COMPLEMENTARY_HARD_DEPENDS_ONLY
SDKIMAGE_INSTALL_COMPLEMENTARY_HARD_DEPENDS_ONLY is a new variable to
enable installation of recommended complementary packages in the classic
sdk. Use it to get populate_sdk behave like Kirkstone and earlier.

As installing complementary recommended packages is known to cause bugs
it was disabled. Though it can be usefull when having dummy packages
with DEPENDS to create SDK's. Use this variable with caution.

Signed-off-by: Claus Stovgaard <claus.stovgaard@gmail.com>